requirements.txt
altair==5.2.0
attrs==23.2.0
blinker==1.7.0
blis==0.7.11
cachetools==5.3.2
catalogue==1.0.2
certifi==2024.2.2
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.2
cymem==2.0.8
docx2txt==0.8
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-2.3.1/en_core_web_sm-2.3.1.tar.gz
filelock==3.13.1
flashtext==2.7
fsspec==2024.2.0
gitdb==4.0.11
GitPython==3.1.41
html2text==2020.1.16
huggingface-hub==0.20.3
idna==3.6
importlib-metadata==7.0.1
Jinja2==3.1.3
joblib==1.3.2
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
lxml==5.1.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.4
packaging==23.2
pandas==2.2.0
pdfminer.six==20231228
pillow==10.2.0
pinecone-client==3.0.2
plac==1.1.3
preshed==3.0.9
protobuf==4.25.2
pyarrow==15.0.0
pycparser==2.21
pycryptodome==3.20.0
pydeck==0.8.0
Pygments==2.17.2
PyPDF2==3.0.1
pyresparser==1.0.6
pyrsistent==0.20.0
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
rich==13.7.0
rpds-py==0.17.1
safetensors==0.4.2
scikit-learn==1.4.0
scipy==1.12.0
sentence-transformers==2.3.1
sentencepiece==0.1.99
six==1.16.0
smmap==5.0.1
sockit==0.3.1
sortedcontainers==2.4.0
spacy==2.3.9
srsly==1.0.7
streamlit==1.31.0
sympy==1.12
tenacity==8.2.3
thinc==7.4.6
threadpoolctl==3.2.0
tokenizers==0.15.1
toml==0.10.2
toolz==0.12.1
torch==2.2.0
tornado==6.4
tqdm==4.66.1
transformers==4.37.2
typing_extensions==4.9.0
tzdata==2023.4
tzlocal==5.2
urllib3==2.2.0
validators==0.22.0
wasabi==0.10.1
wordtrie==0.0.4
zipp==3.17.0
gdown
